@@ -1,7 +1,6 @@
-from __future__ import absolute_import, division, unicode_literals
 
 
-class Filter(object):
+class Filter:
     def __init__(self, source):
         self.source = source
 
